[CHORE] AI big refactor
