+++ .env.example
@@ -1,0 +1,9 @@
+# This env file belongs here, in the repo root. 
+# Copy this file to be .env and fill in the missing values
+# for the blanked out values, reach out to Nick
+POSTGRES_DB=mydb
+POSTGRES_USER=user
+POSTGRES_PASSWORD=pass
+
+DISCORD_CLIENT_ID=
+DISCORD_CLIENT_SECRET=
