+++ apps/backend/ensembles/views/__init__.py
@@ -1,0 +1,17 @@
+from ensembles.views.arrangement import (
+    ArrangementByIdViewSet,
+    ArrangementViewSet,
+    BaseArrangementViewSet,
+)
+from ensembles.views.arrangement_version import ArrangementVersionViewSet
+from ensembles.views.ensemble import EnsembleViewSet
+from ensembles.views.join import JoinEnsembleView
+
+__all__ = [
+    "ArrangementByIdViewSet",
+    "ArrangementVersionViewSet",
+    "ArrangementViewSet",
+    "BaseArrangementViewSet",
+    "EnsembleViewSet",
+    "JoinEnsembleView",
+]
