+++ apps/frontend/.env.example
@@ -1,0 +1,3 @@
+# Vite environment variables (copy to `.env.development` or `.env.local` as needed).
+VITE_API_URL=http://localhost:8000/api
+VITE_FE_URL=http://localhost:5173
