+++ apps/frontend/src/App.tsx
@@ -1,50 +1,7 @@
-import { Routes, Route } from 'react-router-dom';
-import Layout from './Layout';
-import LandingPage from './pages/LandingPage';
-import NotFound from './pages/NotFound';
-import EnsemblesPage from './pages/ensembles/EnsemblesPage';
-import CreateEnsemblePage from './pages/ensembles/CreateEnsemblePage'
-import CreateArrangementPage from './pages/ensembles/CreateArrangementPage'
-import ArrangementsPage from './pages/ensembles/EnsembleArrangementsPage';
-import UploadArrangementVersionFromCommitPage from './pages/ensembles/UploadArrangementVersionAsCommitPage';
-import UploadArrangementVersionDirectlyPage from './pages/ensembles/UploadArrangementVersionDirectlyPage';
-import CreateVersionFromCommitPage from './pages/ensembles/CreateVersionFromCommitPage';
-import ArrangementDisplay from './pages/ensembles/ArrangementDetailPage';
-import {UploadPartsForm} from './pages/testing/UploadPartsForm'
-import PartFormatterPage from './pages/PartFormatterPage'
-import HomePage from './pages/ensembles/HomePage';
-import JoinEnsemblePage from './pages/ensembles/JoinEnsemblePage';
-import EnsembleDisplay from './pages/ensembles/EnsembleDetailPage';
+import { AppRoutes } from "./routes";
 
 function App() {
-  return (
-    <Routes>
-      <Route element={<Layout />}>
-        <Route path="/" element={<LandingPage />} />
-        <Route path="/join/:token" element={<JoinEnsemblePage />} />
-        <Route path="/app" element={<HomePage />} />
-        <Route path="/app/ensembles" element={<EnsemblesPage />} />
-        <Route path="/app/ensembles/create" element={<CreateEnsemblePage />} />
-        <Route path="/app/ensembles/:slug" element={<EnsembleDisplay />} />
-        <Route path="/app/ensembles/:slug/arrangements" element={<ArrangementsPage />} />
-        <Route path="/app/ensembles/:slug/create-arrangement" element={<CreateArrangementPage />} />
-        {/* Future routes */}
-        <Route path="/app/arrangements/:arrangementId" element={<ArrangementDisplay />} />
-        {/* route: /arangements/:slug/v/(latest || versionNum) should get a specific version */}
-        <Route path="/app/arrangements/:slug/edit" element={<div>Edit Arrangement Page (Coming Soon)</div>} />
-        <Route path="/app/arrangements/:arrangementId/new-commit" element={<UploadArrangementVersionFromCommitPage />} />
-        <Route path="/app/arrangements/:arrangementId/new-version" element={<UploadArrangementVersionDirectlyPage />} />
-        <Route
-          path="/app/arrangements/:arrangementId/commits/:commitId/create-version"
-          element={<CreateVersionFromCommitPage />}
-        />
-        <Route path="/testing/upload-arrangement" element={<UploadPartsForm />} />
-
-        <Route path="/part-formatter" element={<PartFormatterPage/>} />
-      </Route>
-      <Route path="*" element={<NotFound />} />
-    </Routes>
-  );
+  return <AppRoutes />;
 }
 
-export default App
+export default App;
