+++ apps/frontend/src/api/index.ts
@@ -1,0 +1,11 @@
+import { arrangementApi } from "./arrangements";
+import { authApi } from "./auth";
+import { ensembleApi } from "./ensembles";
+
+export * from "./types";
+
+export const apiService = {
+  ...authApi,
+  ...ensembleApi,
+  ...arrangementApi,
+};
