+++ pyproject.toml
@@ -1,0 +1,25 @@
+[tool.ruff]
+target-version = "py311"
+line-length = 120
+exclude = [
+  "**/migrations/**",
+  "**/__pycache__",
+  "**/.venv",
+  "**/venv",
+  "**/blob/**",
+  "**/*.egg-info",
+]
+
+[tool.ruff.lint]
+select = ["E4", "E7", "E9", "F", "I", "UP006", "UP007"]
+ignore = [
+  "F403",
+  "F405",
+]
+
+[tool.ruff.lint.per-file-ignores]
+"apps/backend/backend/__init__.py" = ["F401"]
+"**/migrations/**" = ["F401", "I001"]
+
+[tool.ruff.lint.isort]
+known-first-party = ["backend", "core", "divisi", "ensembles"]
